requirements.txt
aiohttp==3.13.3
ib_async==2.1.0
tzdata==2025.3
python-dotenv==1.2.2
